+++ NEWS.md
@@ -1,3 +1,16 @@
+# CoxAIPW 0.0.2
+
+- CoxAIPW function now also outputs cumulative baseline hazard function and survival function for both groups, all of which evaluated at each follow up time.
+
+- CoxAIPW function now supports an additional 'random_censoring' parameter, which allows simplified estimation under the random censoring assumption instead of the informative censoring assumption.
+
+- CoxAIPW function now supports data with tied event times.
+
+- Update referenced articles in the DESCRIPTION.
+
+---
+
+
 # CoxAIPW 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.